New UUIDConverter consistent with other converters.
